requirements.txt
aiogram==3.4.1
openai==1.40.2
python-dotenv==1.0.1
Pillow==10.4.0
